Divide umbral.fragments into umbral.kfrags and umbral.cfrags
